backend/requirements.txt
fastapi
uvicorn
google-genai
pydantic
gitpython
python-dotenv